requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
nltk
ipykernel
transformers
lime
--extra-index-url https://download.pytorch.org/whl/cu126
torch
torchvision
torchaudio
fastapi
uvicorn
pydantic
python-multipart
python-dotenv